entity작업

+++ src/main/java/com/cyworld/cyworldProject/dto/GuestDto.java
@@ -11,9 +11,9 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @AllArgsConstructor
 public class GuestDto {
-    private int id;
-    private int user_id;
-    private int author_id;
+    private long id;
+    private long user_id;
+    private long author_id;
     private String message;
     private Timestamp wdate;
     private int visible;
